Reorganize and rename pages.
