+++ app/assets/javascripts/components/pages/Error/ErrorPage.jsx
