+++ app/views/collections/index.html.erb
@@ -1,1 +1,1 @@
-<%= react_component('CollectionsListPage', collections: @collections_url) %>
+<%= react_component('SiteIndex', collections: @collections_url) %>
