+++ app/views/collections/intro.html.erb
@@ -1,1 +1,1 @@
-<%= react_component('CollectionIntroPage', collection: @collections_url) %>
+<%= react_component('CollectionIntroduction', collection: @collections_url) %>
