+++ app/views/collections/show.html.erb
@@ -1,1 +1,1 @@
-<%= react_component('CollectionShowPage', collection: @collections_url) %>
+<%= react_component('Collection', collection: @collections_url) %>
