+++ app/views/items/search.html.erb
@@ -1,1 +1,1 @@
-<%= react_component('ItemsSearchPage', {collection: @collections_url, hits: @hit_url, searchTerm: params[:q], sortTerm: params[:sort], facet: params[:facet]}, start: params[:start]) %>
+<%= react_component('Search', {collection: @collections_url, hits: @hit_url, searchTerm: params[:q], sortTerm: params[:sort], facet: params[:facet]}, start: params[:start]) %>
