+++ app/views/showcases/show.html.erb
@@ -1,1 +1,1 @@
-<%= react_component('ShowcaseShowPage', collection: @showcases_url ) %>
+<%= react_component('Showcase', collection: @showcases_url ) %>
